avances varios

+++ src/components/carts/routes.js
@@ -32,33 +32,47 @@ router.post("/", async (req, res, next) => {
 });
 
 //*Buscar carrito por id y mostrarlo
-router.get("/:cid", async (req, res, next) => {
+router.get("/:cid", authToken, async (req, res, next) => {
   try {
-    const result = await controller.seeOneCart(req);
-    if (result.status == 200) {
-      return res.status(200).send(result.payload.response);
+    if (req.params.cid == req.user.cart) {
+      const result = await controller.seeOneCart(req);
+      if (result.status == 200) {
+        return res.status(200).send(result.payload.response);
+      }
+      return res.status(400).send(result.payload.result);
+    } else {
+      res.status(401).send({
+        status: 401,
+        response: "El carrito no corresponde al token enviado",
+      });
     }
-    return res.status(400).send(result.payload.result);
   } catch (error) {
     next(error);
   }
 });
 
 //* Agregar producto a carrito
-router.post("/:cid/products/:pid", async (req, res, next) => {
+router.post("/:cid/products/:pid", authToken, async (req, res, next) => {
   try {
-    const result = await controller.addProductToCart(req);
-    if (result.status == 200) {
-      return res.status(200).send(result.payload.response);
+    if (req.params.cid == req.user.cart) {
+      const result = await controller.addProductToCart(req);
+      if (result.status == 200) {
+        return res.status(200).send(result.payload.response);
+      }
+      return res.status(400).send(result.payload.result);
+    } else {
+      res.status(401).send({
+        status: 401,
+        response: "El carrito no corresponde al token enviado",
+      });
     }
-    return res.status(400).send(result.payload.result);
   } catch (error) {
     next(error);
   }
 });
 
-//*Eliminar carrito
-router.delete("/:cid/products/:pid", async (req, res, next) => {
+//*Eliminar producto carrito
+router.delete("/:cid/products/:pid", authToken, async (req, res, next) => {
   try {
     const result = await controller.deleteProduct(req);
     if (result.status == 200) {
@@ -71,13 +85,20 @@ router.delete("/:cid/products/:pid", async (req, res, next) => {
 });
 
 //* Actualizar el carrito con un arreglo
-router.put("/:cid", async (req, res, next) => {
+router.put("/:cid", authToken, async (req, res, next) => {
   try {
-    const result = await controller.updateWithArray(req);
-    if (result.status == 200) {
-      return res.status(200).send(result.payload.result);
+    if (req.params.cid == req.user.cart) {
+      const result = await controller.updateWithArray(req);
+      if (result.status == 200) {
+        return res.status(200).send(result.payload);
+      }
+      return res.status(400).send(result.payload.result);
+    } else {
+      res.status(401).send({
+        status: 401,
+        response: "El carrito no corresponde al token enviado",
+      });
     }
-    return res.status(400).send(result.payload.result);
   } catch (error) {
     next(error);
   }
@@ -97,13 +118,20 @@ router.put("/:cid/products/:pid", async (req, res, next) => {
 });
 
 //* Eliminar todos los productos de un carrito
-router.delete("/:cid", async (req, res, next) => {
+router.delete("/:cid", authToken, async (req, res, next) => {
   try {
-    const result = await controller.deleteAllProducts(req);
-    if (result.status == 200) {
-      return res.status(200).send(result.payload.result);
+    if (req.params.cid == req.user.cart) {
+      const result = await controller.deleteAllProducts(req);
+      if (result.status == 200) {
+        return res.status(200).send(result.payload.result);
+      }
+      return res.status(400).send(result.payload.result);
+    } else {
+      res.status(401).send({
+        status: 401,
+        response: "El carrito no corresponde al token enviado",
+      });
     }
-    return res.status(400).send(result.payload.result);
   } catch (error) {
     next(error);
   }
@@ -114,9 +142,13 @@ router.post("/:cid/purchase", authToken, async (req, res, next) => {
   try {
     if (req.params.cid == req.user.cart) {
       let user = new PurchaseDTO(req.user, req.headers.token);
+     
       const result = await controller.buyProducts(user);
       if (result.status == 200) {
-        return res.status(result.status).send(result.response);
+        return res.status(result.status).send({
+          status: result.status,
+          response: result.response,
+        });
       } else {
         return res.status(result.status).send({
           status: result.status,
